feat: explicitly set client timeout
problem: running into issues where the client is timing out.  But the
issue is difficult to debug as it is only being hit in the wild with
upload times over a minute.  The client also claims there is not a
cleitn defined timeout if one isn't provided.  so it is confusing

solution: explicitly set long timeout on the client.